+++ src/fr/imac/javawars/engine/Unite.java
@@ -7,12 +7,12 @@ import fr.imac.javawars.player.Player;
 public interface Unite {
 	
 	//methods
-	public void loseLife(int loseLife);
-	public void addLife(int addLife);
+	public void loseLife(float loseLife);
+	public void addLife(float addLife);
 	
 	//getters/setters
-	public void setLife(int life);
-	public int getLife();
+	public void setLife(float life);
+	public float getLife();
 	public void setPosition(Point p);
 	public Point getPosition();
 	public Player getPlayer();
